requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-database-url==1.2.0
dj-email-url==1.0.6
Django==4.1.3
django-allauth==0.51.0
django-cache-url==3.4.4
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-widget-tweaks==1.4.12
environs==9.5.0
idna==3.4
marshmallow==3.19.0
oauthlib==3.2.2
packaging==22.0
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
python-gettext==4.1
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
